feat: toggle story list with redux state
